requirements.txt
# Core dependencies
python-docx>=0.8.11
langchain>=0.1.0
pydantic>=2.0.0

# LLM providers
openai>=1.0.0 # Added dependency
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
langchain-anthropic>=0.0.5

# Text processing
regex>=2023.0.0

# Testing
pytest>=7.0.0

# Graph & Embeddings
neo4j>=5.10.0 # Check for latest compatible version
sentence-transformers>=2.2.0 # Check for latest compatible version
scikit-learn>=1.0 # For cosine similarity, check latest compatible version

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
PyYAML>=6.0 # Added for reading experience.yml
